pyproject.toml
[tool.black]
line-length = 120

[tool.poetry]
name = "simple-stocks"
version = "0.1.0"
description = ""
authors = ["MarcDufresne <marc.andre.dufresne@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Sanic = "^0.8.3"
ujson = "^1.35"
aiohttp = "^3.3"
jinja2 = "^2.10"
sanic-jinja2 = "^0.7.1"
sanic_session = "^0.4.1"
sanic-jwt = "^1.2"
sanic_oauth = "^0.2.5"
asyncio_redis = "^0.15.1"
psycopg2 = "^2.7"
pysqlite3 = "^0.2.0"
peewee-migrate = "^1.1"
munch = "^2.3"
click = "^7.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.6"
pytest = "^3.10"
black = "^18.3-alpha.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
